Implement common tests in base class
